backend/singleview/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
django-requestlogs==0.2.3
django-ipware==3.0.2
python-dateutil==2.8.1